+++ app/src/main/java/com/example/healthappttt/Fragment/HomeFragmentBackUp.java
@@ -155,7 +155,7 @@ public class HomeFragmentBackUp extends Fragment {
 
                                 // Log.d(TAG, document.getId() + " &&&&+&=> " + document.getData().get("userName").toString());
                                 User a= new User(
-                                        Double.parseDouble(document.getData().get("userTemperature").toString()),
+                                        (int) Double.parseDouble(document.getData().get("userTemperature").toString()),
                                         document.getData().get("key_").toString(),
                                         Double.parseDouble(document.getData().get("lat").toString()),
                                         Double.parseDouble(document.getData().get("lon").toString()),
